leafgen: make generic rewrite value-only
