+++ TreeDB/db/vlog_rewrite_bench_test.go
@@ -12,9 +12,7 @@ import (
 
 	batchpkg "github.com/snissn/gomap/TreeDB/batch"
 	"github.com/snissn/gomap/TreeDB/internal/valuelog"
-	"github.com/snissn/gomap/TreeDB/node"
 	"github.com/snissn/gomap/TreeDB/page"
-	"github.com/snissn/gomap/TreeDB/pager"
 )
 
 func BenchmarkValueLogRewriteOnline_ValuePointers(b *testing.B) {
@@ -59,103 +57,6 @@ func BenchmarkValueLogRewriteOnline_ValuePointers(b *testing.B) {
 	if b.N > 0 {
 		b.ReportMetric(float64(totalCopied)/float64(b.N), "value_records/op")
 		b.ReportMetric(float64(totalBytes)/float64(b.N), "value_bytes/op")
-		b.ReportMetric(float64(totalRefreshScans)/float64(b.N), "refresh_scans/op")
-		b.ReportMetric(float64(totalRewriteAllocs)/float64(b.N), "rewrite_allocs/op")
-	}
-}
-
-func BenchmarkValueLogRewriteOnline_LeafRefs(b *testing.B) {
-	const keyCount = 1536
-
-	var totalCopied int64
-	var totalBytes int64
-	var totalRefreshScans uint64
-	var totalRewriteAllocs uint64
-
-	for i := 0; i < b.N; i++ {
-		b.StopTimer()
-		db, sourceIDs, cleanup := setupLeafRefRewriteBench(b, keyCount)
-		refreshBefore := db.valueLogManager.RefreshScanCount()
-		var memBefore runtime.MemStats
-		runtime.ReadMemStats(&memBefore)
-		b.StartTimer()
-
-		stats, err := db.ValueLogRewriteOnline(context.Background(), ValueLogRewriteOnlineOptions{
-			SourceFileIDs: sourceIDs,
-			BatchSize:     256,
-		})
-		b.StopTimer()
-		if err != nil {
-			cleanup()
-			b.Fatalf("ValueLogRewriteOnline: %v", err)
-		}
-		totalCopied += int64(stats.LeafRefRecordsCopied)
-		totalBytes += stats.LeafRefBytesCopied
-		totalRefreshScans += db.valueLogManager.RefreshScanCount() - refreshBefore
-		var memAfter runtime.MemStats
-		runtime.ReadMemStats(&memAfter)
-		if memAfter.Mallocs > memBefore.Mallocs {
-			totalRewriteAllocs += memAfter.Mallocs - memBefore.Mallocs
-		}
-		cleanup()
-	}
-
-	if b.N > 0 {
-		b.ReportMetric(float64(totalCopied)/float64(b.N), "leafref_records/op")
-		b.ReportMetric(float64(totalBytes)/float64(b.N), "leafref_bytes/op")
-		b.ReportMetric(float64(totalRefreshScans)/float64(b.N), "refresh_scans/op")
-		b.ReportMetric(float64(totalRewriteAllocs)/float64(b.N), "rewrite_allocs/op")
-	}
-}
-
-func BenchmarkValueLogRewriteOnline_LeafRefs_ReserveRIDs(b *testing.B) {
-	const keyCount = 1536
-
-	var totalCopied int64
-	var totalBytes int64
-	var totalRefreshScans uint64
-	var totalRewriteAllocs uint64
-
-	for i := 0; i < b.N; i++ {
-		b.StopTimer()
-		db, sourceIDs, cleanup := setupLeafRefRewriteBench(b, keyCount)
-		refreshBefore := db.valueLogManager.RefreshScanCount()
-		nextRID := uint64(1 << 42)
-		var memBefore runtime.MemStats
-		runtime.ReadMemStats(&memBefore)
-		b.StartTimer()
-
-		stats, err := db.ValueLogRewriteOnline(context.Background(), ValueLogRewriteOnlineOptions{
-			SourceFileIDs: sourceIDs,
-			BatchSize:     256,
-			ReserveRIDs: func(count int) (uint64, error) {
-				if count <= 0 {
-					return 0, fmt.Errorf("invalid count %d", count)
-				}
-				start := nextRID
-				nextRID += uint64(count)
-				return start, nil
-			},
-		})
-		b.StopTimer()
-		if err != nil {
-			cleanup()
-			b.Fatalf("ValueLogRewriteOnline: %v", err)
-		}
-		totalCopied += int64(stats.LeafRefRecordsCopied)
-		totalBytes += stats.LeafRefBytesCopied
-		totalRefreshScans += db.valueLogManager.RefreshScanCount() - refreshBefore
-		var memAfter runtime.MemStats
-		runtime.ReadMemStats(&memAfter)
-		if memAfter.Mallocs > memBefore.Mallocs {
-			totalRewriteAllocs += memAfter.Mallocs - memBefore.Mallocs
-		}
-		cleanup()
-	}
-
-	if b.N > 0 {
-		b.ReportMetric(float64(totalCopied)/float64(b.N), "leafref_records/op")
-		b.ReportMetric(float64(totalBytes)/float64(b.N), "leafref_bytes/op")
 		b.ReportMetric(float64(totalRefreshScans)/float64(b.N), "refresh_scans/op")
 		b.ReportMetric(float64(totalRewriteAllocs)/float64(b.N), "rewrite_allocs/op")
 	}
@@ -238,161 +139,6 @@ func setupValuePointerRewriteBench(tb testing.TB, seg1Records, seg2Records int) 
 		_ = os.RemoveAll(dir)
 	}
 	return db, sourceIDs, cleanup
-}
-
-func setupLeafRefRewriteBench(tb testing.TB, keyCount int) (*DB, []uint32, func()) {
-	tb.Helper()
-	dir, err := os.MkdirTemp("", "treedb-vlog-rewrite-leaf-bench-*")
-	if err != nil {
-		tb.Fatalf("MkdirTemp: %v", err)
-	}
-
-	db, err := Open(Options{
-		Dir:                        dir,
-		Durability:                 DurabilityWALOffRelaxed,
-		DisableBackgroundPrune:     true,
-		IndexOuterLeavesInValueLog: true,
-		LeafPrefixCompression:      true,
-		IndexColumnarLeaves:        true,
-		IndexPackedValuePtr:        true,
-	})
-	if err != nil {
-		_ = os.RemoveAll(dir)
-		tb.Fatalf("Open: %v", err)
-	}
-
-	walDir := filepath.Join(dir, "value_vlog")
-	if err := os.MkdirAll(walDir, 0o755); err != nil {
-		_ = db.Close()
-		_ = os.RemoveAll(dir)
-		tb.Fatalf("mkdir wal: %v", err)
-	}
-
-	leafLog := newRewriteWriter(walDir, 0, 0, 16<<10)
-	leafLog.blockCompression = false
-	leafLog.blockCodec = valuelog.BlockCodecSnappy
-	db.SetLeafPageLog(leafLog)
-
-	value := bytes.Repeat([]byte("leafref-bench-value-"), 3)
-	bt, ok := db.NewBatch().(*Batch)
-	if !ok {
-		_ = leafLog.Close()
-		_ = db.Close()
-		_ = os.RemoveAll(dir)
-		tb.Fatalf("NewBatch type assertion failed")
-	}
-	for i := 0; i < keyCount; i++ {
-		key := []byte(fmt.Sprintf("k%06d", i))
-		if err := bt.Set(key, value); err != nil {
-			_ = bt.Close()
-			_ = leafLog.Close()
-			_ = db.Close()
-			_ = os.RemoveAll(dir)
-			tb.Fatalf("Batch.Set(%q): %v", key, err)
-		}
-	}
-	if err := bt.Write(); err != nil {
-		_ = bt.Close()
-		_ = leafLog.Close()
-		_ = db.Close()
-		_ = os.RemoveAll(dir)
-		tb.Fatalf("seed Write: %v", err)
-	}
-	_ = bt.Close()
-	if err := db.RefreshValueLogSet(); err != nil {
-		_ = leafLog.Close()
-		_ = db.Close()
-		_ = os.RemoveAll(dir)
-		tb.Fatalf("RefreshValueLogSet: %v", err)
-	}
-	state := db.State()
-	if state == nil {
-		_ = leafLog.Close()
-		_ = db.Close()
-		_ = os.RemoveAll(dir)
-		tb.Fatalf("missing DB state")
-	}
-
-	counts := make(map[uint32]int, 16)
-	collectLeafRefFileCountsBench(tb, db.Pager(), state.RootPageID, counts)
-	collectLeafRefFileCountsBench(tb, db.Pager(), state.SystemRootPageID, counts)
-	if len(counts) == 0 {
-		_ = leafLog.Close()
-		_ = db.Close()
-		_ = os.RemoveAll(dir)
-		tb.Fatalf("no leafref source file IDs found")
-	}
-
-	targetID := uint32(0)
-	targetCount := -1
-	for fileID, n := range counts {
-		if n > targetCount || (n == targetCount && fileID < targetID) {
-			targetID = fileID
-			targetCount = n
-		}
-	}
-
-	sourceIDs := []uint32{targetID}
-	cleanup := func() {
-		_ = leafLog.Close()
-		_ = db.Close()
-		_ = os.RemoveAll(dir)
-	}
-	return db, sourceIDs, cleanup
-}
-
-func collectLeafRefFileCountsBench(tb testing.TB, p *pager.Pager, rootID uint64, counts map[uint32]int) {
-	tb.Helper()
-	if p == nil || rootID == 0 || counts == nil {
-		return
-	}
-	if ptr, ok := page.DecodeLeafRef(rootID); ok {
-		counts[ptr.ValueLogFileID()]++
-		return
-	}
-
-	stack := make([]uint64, 0, 128)
-	stack = append(stack, rootID)
-	visited := make(map[uint64]struct{}, 256)
-
-	for len(stack) > 0 {
-		pageID := stack[len(stack)-1]
-		stack = stack[:len(stack)-1]
-		if _, ok := visited[pageID]; ok {
-			continue
-		}
-		visited[pageID] = struct{}{}
-
-		if ptr, ok := page.DecodeLeafRef(pageID); ok {
-			counts[ptr.ValueLogFileID()]++
-			continue
-		}
-
-		data, err := p.Get(pageID)
-		if err != nil {
-			tb.Fatalf("pager.Get(%d): %v", pageID, err)
-		}
-		n := node.NewNodeView(data)
-		if !n.VerifyChecksum() {
-			tb.Fatalf("checksum mismatch on page %d", pageID)
-		}
-
-		switch n.Type() {
-		case page.PageTypeLeaf:
-			// no children
-		case page.PageTypeInternal:
-			count := n.Count()
-			for i := uint16(0); i < count; i++ {
-				childID, err := n.GetInternalChildID(i)
-				if err != nil {
-					tb.Fatalf("GetInternalChildID(%d,%d): %v", pageID, i, err)
-				}
-				stack = append(stack, childID)
-			}
-		default:
-			tb.Fatalf("unexpected page type %d at page %d", n.Type(), pageID)
-		}
-	}
 }
 
 func appendPointersInNewSegmentBench(tb testing.TB, dir string, lane, seq uint32, ridBase uint64, n int, valueAt func(i int) []byte) []page.ValuePtr {
